refactor(reviewer-bot): extract reconcile handlers
